Finish all the buttons on users page
the only button that needs to be finish is the delete button. that however is easily done. after that i will be going to work on the top buttons to go between all the wiki and post pages. after that i will work on the info pages form all the wikis and post pages.